backend/requirements.txt
fastapi
uvicorn
playwright
groq
python-dotenv
psycopg2-binary
asyncpg